+++ include/cpu/ooo/ooo_cpu.h
@@ -53,8 +53,8 @@ public:
     void reset() override;                   // 重置CPU状态
     
     // 寄存器访问（架构寄存器值）
-    uint32_t getRegister(RegNum reg) const override;
-    void setRegister(RegNum reg, uint32_t value) override;
+    uint64_t getRegister(RegNum reg) const override;
+    void setRegister(RegNum reg, uint64_t value) override;
     
     // 浮点寄存器访问
     uint32_t getFPRegister(RegNum reg) const override;
@@ -63,8 +63,8 @@ public:
     void setFPRegisterFloat(RegNum reg, float value) override;
     
     // 程序计数器
-    uint32_t getPC() const override { return cpu_state_.pc; }
-    void setPC(uint32_t pc) override { cpu_state_.pc = pc; }
+    uint64_t getPC() const override { return cpu_state_.pc; }
+    void setPC(uint64_t pc) override { cpu_state_.pc = pc; }
     
     // 状态查询
     bool isHalted() const override { return cpu_state_.halted; }
@@ -88,7 +88,7 @@ public:
     void setDiffTest(class DiffTest* difftest) override;  // 由Simulator设置DiffTest引用
     void enableDiffTest(bool enable);
     bool isDiffTestEnabled() const override;
-    void performDiffTestWithCommittedPC(uint32_t committed_pc) override;
+    void performDiffTestWithCommittedPC(uint64_t committed_pc) override;
     void getDiffTestStats(uint64_t& comparisons, uint64_t& mismatches) const;
     
 private:
@@ -112,26 +112,26 @@ private:
     void print_stage_activity(const std::string& stage, const std::string& activity);
     
     // 异常处理
-    void handle_exception(const std::string& exception_msg, uint32_t pc);
+    void handle_exception(const std::string& exception_msg, uint64_t pc);
     void flush_pipeline();
     
     // 分支预测（简化实现）
-    bool predict_branch(uint32_t pc);
-    void update_branch_predictor(uint32_t pc, bool taken);
+    bool predict_branch(uint64_t pc);
+    void update_branch_predictor(uint64_t pc, bool taken);
     
     // 系统调用处理
     void handleEcall();
     void handleEbreak();
     
     // 辅助方法
-    uint32_t get_physical_register_value(PhysRegNum reg) const;
-    void set_physical_register_value(PhysRegNum reg, uint32_t value);
+    uint64_t get_physical_register_value(PhysRegNum reg) const;
+    void set_physical_register_value(PhysRegNum reg, uint64_t value);
     uint32_t get_physical_fp_register_value(PhysRegNum reg) const;
     void set_physical_fp_register_value(PhysRegNum reg, uint32_t value);
     
     // 内存访问辅助方法
-    uint32_t loadFromMemory(Address addr, Funct3 funct3);
-    void storeToMemory(Address addr, uint32_t value, Funct3 funct3);
+    uint64_t loadFromMemory(Address addr, Funct3 funct3);
+    void storeToMemory(Address addr, uint64_t value, Funct3 funct3);
     
     // 立即数符号扩展
     int32_t signExtend(uint32_t value, int bits) const;
